FastAPI/requirements.txt
cli-helpers==2.2.1
click==8.1.3
configobj==5.0.6
fastapi==0.68.2
h11==0.13.0
humanize==4.1.0
pgcli==1.9.0
pgspecial==2.0.0
prompt-toolkit==1.0.18
psycopg==3.0.14
psycopg2-binary==2.9.3
pydantic==1.9.1
Pygments==2.12.0
setproctitle==1.2.3
six==1.16.0
sqlparse==0.2.4
starlette==0.14.2
tabulate==0.8.9
typing_extensions==4.2.0
uvicorn==0.15.0
wcwidth==0.2.5
